extracting the config models to their own files
