+++ vector-config/src/main/java/im/vector/app/config/Analytics.kt
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2022 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package im.vector.app.config
+
+/**
+ * The types of analytics Element currently supports.
+ */
+sealed interface Analytics {
+
+    /**
+     * Disables the analytics integrations.
+     */
+    object Disabled : Analytics
+
+    /**
+     * Analytics integration via PostHog.
+     */
+    data class PostHog(
+            /**
+             * The PostHog instance url.
+             */
+            val postHogHost: String,
+
+            /**
+             * The PostHog instance API key.
+             */
+            val postHogApiKey: String,
+
+            /**
+             * A URL to more information about the analytics collection.
+             */
+            val policyLink: String,
+    ) : Analytics
+}
