+++ vector-config/src/main/java/im/vector/app/config/Config.kt
@@ -87,58 +87,3 @@ object Config {
      */
     val NIGHTLY_ANALYTICS_CONFIG = RELEASE_ANALYTICS_CONFIG
 }
-
-/**
- * The types of analytics Element currently supports.
- */
-sealed interface Analytics {
-
-    /**
-     * Disables the analytics integrations.
-     */
-    object Disabled : Analytics
-
-    /**
-     * Analytics integration via PostHog.
-     */
-    data class PostHog(
-            /**
-             * The PostHog instance url.
-             */
-            val postHogHost: String,
-
-            /**
-             * The PostHog instance API key.
-             */
-            val postHogApiKey: String,
-
-            /**
-             * A URL to more information about the analytics collection.
-             */
-            val policyLink: String,
-    ) : Analytics
-}
-
-enum class OnboardingVariant {
-    LEGACY,
-    LOGIN_2,
-    FTUE_AUTH
-}
-
-enum class KeySharingStrategy {
-    /**
-     * Keys will be sent for the first time when the first message is sent.
-     * This is handled by the Matrix SDK so there's no need to do it in Vector.
-     */
-    WhenSendingEvent,
-
-    /**
-     * Keys will be sent for the first time when the timeline displayed.
-     */
-    WhenEnteringRoom,
-
-    /**
-     * Keys will be sent for the first time when a typing started.
-     */
-    WhenTyping
-}
